requirements.txt
python = 3.9
pdfminer.six
python-docx
flask
spacy